refactor(api): mv rna-siren in open-data
