+++ packages/api/src/modules/providers/chorus/chorus.service.ts
@@ -4,7 +4,7 @@ import CacheData from "../../../shared/Cache";
 import { asyncFilter } from "../../../shared/helpers/ArrayHelper";
 import { siretToSiren } from "../../../shared/helpers/SirenHelper";
 import { isEJ, isSiret } from "../../../shared/Validators";
-import rnaSirenService from "../../rna-siren/rnaSiren.service";
+import rnaSirenService from "../../open-data/rna-siren/rnaSiren.service";
 import Versement from "../../versements/@types/Versement";
 import VersementsProvider from "../../versements/@types/VersementsProvider";
 import dataGouvService from "../datagouv/datagouv.service";
