+++ packages/api/tests/modules/open-data/rna-siren/rna-siren.service.test.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import Association from "../../../src/modules/associations/@types/Association";
-import leCompteAssoService from "../../../src/modules/providers/leCompteAsso/leCompteAsso.service";
-import osirisService from "../../../src/modules/providers/osiris/osiris.service";
+import Association from "../../../../src/modules/associations/@types/Association";
+import leCompteAssoService from "../../../../src/modules/providers/leCompteAsso/leCompteAsso.service";
+import osirisService from "../../../../src/modules/providers/osiris/osiris.service";
 import RnaSiren from "../../../src/modules/rna-siren/entities/RnaSirenEntity";
 import rnaSirenService from "../../../src/modules/rna-siren/rnaSiren.service"
-import db from "../../../src/shared/MongoConnection";
+import db from "../../../../src/shared/MongoConnection";
 
 // Not the same as entity from MongoDB
 // It is supposed to reference new RnaSiren from a provider document
